+++ thermion_dart/lib/src/filament/filament.dart
@@ -13,6 +13,7 @@ export 'src/interface/tone_mapper.dart';
 export 'src/interface/gltf_mesh_data.dart';
 export 'src/interface/vertex_buffer.dart';
 export 'src/interface/index_buffer.dart';
+export 'src/interface/buffer_object.dart';
 export 'src/interface/translation_axis_material.dart';
 export 'src/interface/ubershader_material.dart';
 export 'src/interface/wireframe_material.dart';