+++ thermion_dart/native/include/c_api/ThermionDartRenderThreadApi.h
@@ -8,6 +8,7 @@
 #include "TMaterialProvider.h"
 #include "TVertexBuffer.h"
 #include "TIndexBuffer.h"
+#include "TBufferObject.h"
 #include "TTransformManager.h"
 #include "TLightManager.h"
 
@@ -405,6 +406,35 @@ namespace thermion
             void* data,
             size_t sizeInBytes,
             uint32_t byteOffset,
+            uint32_t requestId,
+            VoidCallback onComplete
+        );
+        EMSCRIPTEN_KEEPALIVE void VertexBuffer_setBufferObjectAtRenderThread(
+            TEngine* tEngine,
+            TVertexBuffer* tBuffer,
+            uint8_t bufferIndex,
+            TBufferObject* tBufferObject,
+            uint32_t requestId,
+            VoidCallback onComplete
+        );
+
+        EMSCRIPTEN_KEEPALIVE void BufferObjectBuilder_buildRenderThread(
+            TBufferObjectBuilder* tBuilder,
+            TEngine* tEngine,
+            void (*onComplete)(TBufferObject*)
+        );
+        EMSCRIPTEN_KEEPALIVE void BufferObject_setBufferRenderThread(
+            TEngine* tEngine,
+            TBufferObject* tBuffer,
+            void* data,
+            size_t sizeInBytes,
+            uint32_t byteOffset,
+            uint32_t requestId,
+            VoidCallback onComplete
+        );
+        EMSCRIPTEN_KEEPALIVE void BufferObject_destroyRenderThread(
+            TEngine* tEngine,
+            TBufferObject* tBuffer,
             uint32_t requestId,
             VoidCallback onComplete
         );